+++ libs/platform/experience/src/services/data-client/data-client.model.ts
@@ -1,7 +1,7 @@
 import { ModeEvent } from '@spryker-oryx/ui/color-mode-selector';
-import { ContentComponentSchema } from '../../../models';
-import { ResourceGraphic } from '../../../plugins';
-import { StaticComponent } from '../static-data';
+import { ContentComponentSchema } from '../../models';
+import { ResourceGraphic } from '../../plugins';
+import { StaticComponent } from '../experience/static-data';
 
 export const enum MessageType {
   Graphics = 'oryx.graphics',
