+++ libs/platform/experience/src/services/data-client/default-data-client.service.spec.ts
@@ -7,12 +7,12 @@ import {
 } from '@spryker-oryx/core';
 import { createInjector, destroyInjector, getInjector } from '@spryker-oryx/di';
 import { of } from 'rxjs';
-import { optionsKey } from '../../../decorators';
-import { ResourcePlugin } from '../../../plugins';
-import { postMessage } from '../utilities';
+import { optionsKey } from '../../decorators';
+import { ResourcePlugin } from '../../plugins';
 import { MessageType } from './data-client.model';
 import { ExperienceDataClientService } from './data-client.service';
 import { DefaultExperienceDataClientService } from './default-data-client.service';
+import { postMessage } from './utilities';
 
 const mockAppFn = {
   getGraphics: vi.fn(),
