+++ libs/platform/experience/src/services/data-client/index.ts
@@ -1,3 +1,4 @@
 export * from './data-client.model';
 export * from './data-client.service';
 export * from './default-data-client.service';
+export * from './utilities';
