+++ libs/platform/experience/src/services/experience/index.ts
@@ -1,7 +1,5 @@
-export * from './data-client';
 export * from './default-experience.service';
 export * from './experience.service';
 export * from './models';
 export * from './preview-experience.service';
 export * from './static-data';
-export * from './utilities';
