+++ libs/platform/experience/src/services/index.ts
@@ -1,5 +1,7 @@
+export * from './data-client';
 export * from './experience';
 export * from './experience-tokens';
 export * from './experience.providers';
 export * from './layout';
 export * from './registry';
+export * from './resolvers';
